+++ packages/core/calendar/src/styled/__tests__/Date.js
@@ -5,12 +5,12 @@ import React from 'react';
 import { DateDiv } from '../../styled/Date';
 
 test('cursor should be "default"', () => {
-  // $FlowFixMe - invalid intersection error.
+  // $FlowFixMe - https://github.com/facebook/flow/issues/396
   expect(shallow(<DateDiv />)).toHaveStyleRule('cursor', 'pointer');
 });
 
 test('disabled - cursor should be "not-allowed"', () => {
-  // $FlowFixMe - invalid intersection error.
+  // $FlowFixMe - https://github.com/facebook/flow/issues/396
   expect(shallow(<DateDiv disabled />)).toHaveStyleRule(
     'cursor',
     'not-allowed',