+++ packages/core/spinner/src/Spinner/styledSvg.js
@@ -84,8 +84,9 @@ export const getStrokeColor = ({
   ...props
 }: {
   invertColor?: boolean,
-  // $FlowFixMe TEMPORARY
-}): string => (invertColor ? spinnerColorInverted(props) : spinnerColor(props));
+}): string | number =>
+  // $FlowFixMe - theme is not found in props
+  invertColor ? spinnerColorInverted(props) : spinnerColor(props);
 
 export const svgStyles = css`
   ${(props: StyleParams) => {