+++ packages/editor/renderer/src/text/nodes/index.ts
@@ -2,6 +2,7 @@ import { Node as PMNode, Schema } from 'prosemirror-model';
 import applicationCard from './applicationCard';
 import blockquote from './blockquote';
 import bulletList from './bulletList';
+import decisionItem from './decisionItem';
 import hardBreak from './hardBreak';
 import heading from './heading';
 import listItem from './listItem';
@@ -11,6 +12,7 @@ import panel from './panel';
 import paragraph from './paragraph';
 import rule from './rule';
 import table from './table';
+import taskItem from './taskItem';
 import unknown from './unknown';
 
 export type NodeReducer = (node: PMNode, schema: Schema) => string;
@@ -25,6 +27,7 @@ export const nodeToReducerMapping: { [key: string]: NodeReducer } = {
   applicationCard,
   blockquote,
   bulletList,
+  decisionItem,
   hardBreak,
   heading,
   listItem,
@@ -34,5 +37,6 @@ export const nodeToReducerMapping: { [key: string]: NodeReducer } = {
   paragraph,
   rule,
   table,
+  taskItem,
   unknown,
 };