+++ packages/core/navigation/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @atlaskit/navigation
 
+## 32.2.3
+- [patch] Fix $FlowFixMe and release packages [25d0b2d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/25d0b2d)
+- [none] Updated dependencies [25d0b2d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/25d0b2d)
+  - @atlaskit/tooltip@10.3.1
+  - @atlaskit/single-select@5.2.1
+  - @atlaskit/button@8.2.2
+  - @atlaskit/analytics@3.0.6
+  - @atlaskit/checkbox@3.1.2
+  - @atlaskit/icon@12.3.1
+  - @atlaskit/avatar@11.2.1
+
 ## 32.2.2
 - [patch] Replaces implementation of ScrollLock with [react-scrolllock](https://github.com/jossmac/react-scrolllock). Deprecates ScrollLock export in @atlaskit/layer-manager. [497d50d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/497d50d)
 - [none] Updated dependencies [497d50d](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/497d50d)